+++ CHANGES.md
@@ -4,7 +4,7 @@ moodle-theme_boost_campus
 Changes
 -------
 
-### Unreleased
+### Release v3.8-r1
 
 * 2020-03-24 - Changed setting imageareaitemslink to imageareaitemsattributes to be able to add alt text to the images.
 * 2020-03-10 - Fixed accessibility issues with in-course settings.